#5/refactor: get Rid off old calculator imlementation. Left the one with the functions.
